+++ jdaviz/configs/imviz/helper.py
@@ -3,8 +3,8 @@ import re
 import warnings
 from copy import deepcopy
 
-from astropy.utils import deprecated
 import numpy as np
+from astropy.utils import deprecated
 from glue.core.link_helpers import LinkSame
 
 from jdaviz.core.events import SnackbarMessage, NewViewerMessage
@@ -313,6 +313,7 @@ class Imviz(ImageConfigHelper):
         """
         return self.plugins['Aperture Photometry'].export_table()
 
+    @deprecated(since="4.2", alternative="plugins['Catalog Search'].export_table()")
     def get_catalog_source_results(self):
         """Return table of sources given by querying from a catalog, if any.
         Results are calculated using :ref:`imviz-catalogs` plugin.