+++ jdaviz/configs/imviz/tests/test_regions.py
@@ -1,3 +1,5 @@
+import warnings
+
 import numpy as np
 import pytest
 from astropy import units as u
@@ -29,7 +31,7 @@ class BaseRegionHandler:
 class TestLoadRegions(BaseImviz_WCS_NoWCS, BaseRegionHandler):
     def teardown_method(self, method):
         """Clear all the subsets for the next test method."""
-        self.imviz._delete_all_regions()
+        self.imviz.app.delete_subsets()
 
     def test_regions_invalid(self):
         # Wrong object
@@ -54,7 +56,7 @@ class TestLoadRegions(BaseImviz_WCS_NoWCS, BaseRegionHandler):
                                                        return_bad_regions=True)
         assert len(bad_regions) == 1 and bad_regions[0][1] == 'Sky region provided but data has no valid WCS'  # noqa
 
-        # Unsupported functionality from outside load_regions
+        # Unsupported functionality
         reg = PointSkyRegion(center=sky)
         bad_regions = self.subset_plugin.import_region(reg, return_bad_regions=True)
         assert len(bad_regions) == 1 and bad_regions[0][1] == 'Failed to load: NotImplementedError()'  # noqa
@@ -86,16 +88,22 @@ class TestLoadRegions(BaseImviz_WCS_NoWCS, BaseRegionHandler):
         bad_regions = self.subset_plugin.import_region([mask], return_bad_regions=True)
         assert len(bad_regions) == 0
         self.verify_region_loaded('MaskedSubset 1')
-        assert self.imviz.plugins['Subset Tools'].get_regions() == {}
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore',
+                                    message='Regions skipped: MaskedSubset 1')
+            assert self.imviz.plugins['Subset Tools'].get_regions() == {}
 
         mask[1, 1] = True
         bad_regions = self.subset_plugin.import_region([mask], return_bad_regions=True)
         assert len(bad_regions) == 0
         self.verify_region_loaded('MaskedSubset 2')
-        assert self.imviz.plugins['Subset Tools'].get_regions() == {}
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore',
+                                    message='Regions skipped: MaskedSubset 1')
+            assert self.imviz.plugins['Subset Tools'].get_regions() == {}
 
         # Also test deletion by label here.
-        self.imviz._delete_region('MaskedSubset 1')
+        self.imviz.app.delete_subsets('MaskedSubset 1')
         self.verify_region_loaded('MaskedSubset 1', count=0)
 
         # Adding another mask will increment from 2 even when 1 is now available.
@@ -103,10 +111,14 @@ class TestLoadRegions(BaseImviz_WCS_NoWCS, BaseRegionHandler):
         bad_regions = self.subset_plugin.import_region([mask], return_bad_regions=True)
         assert len(bad_regions) == 0
         self.verify_region_loaded('MaskedSubset 3')
-        assert self.imviz.plugins['Subset Tools'].get_regions() == {}
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore',
+                                    message='Regions skipped: MaskedSubset 2, MaskedSubset 3')
+            assert self.imviz.plugins['Subset Tools'].get_regions() == {}
 
-        # Deletion of non-existent label is silent no-op.
-        self.imviz._delete_region('foo')
+        # Deletion of non-existent label raises error
+        with pytest.raises(ValueError, match=r'foo not in data collection, can not delete\.'):
+            self.imviz.app.delete_subsets('foo')
 
     def test_regions_pixel(self):
         # A little out-of-bounds should still overlay the overlapped part.
@@ -115,11 +127,15 @@ class TestLoadRegions(BaseImviz_WCS_NoWCS, BaseRegionHandler):
                                                        combination_mode='new')
         assert len(bad_regions) == 0
         self.verify_region_loaded('Subset 1')
-        assert len(self.imviz.plugins['Subset Tools'].get_regions()) == 1
+        st = self.imviz.plugins['Subset Tools']
+        assert len(st.get_regions()) == 1
+        # test passing in the subset label as well
+        assert len(st.get_regions(list_of_subset_labels='Subset 1')) == 1
 
     def test_regions_sky_has_wcs(self):
-        # Mimic interactive region (before)
-        self.imviz._apply_interactive_region('bqplot:truecircle', (1.5, 2.5), (3.6, 4.6))
+        my_reg_pix_1 = CirclePixelRegion(center=PixCoord(x=2.55, y=3.55), radius=1.05)
+        my_reg_pix_2 = EllipsePixelRegion(center=PixCoord(x=1.5, y=2.25), width=7.0, height=4.5)
+        my_reg_pix_3 = RectanglePixelRegion(center=PixCoord(x=5.0, y=5.0), width=10, height=10)
 
         sky = SkyCoord(ra=337.5202808, dec=-20.833333059999998, unit='deg')
         # These will become indistinguishable from normal Subset.
@@ -127,21 +143,21 @@ class TestLoadRegions(BaseImviz_WCS_NoWCS, BaseRegionHandler):
         my_reg_sky_2 = CircleAnnulusSkyRegion(center=sky, inner_radius=0.0004 * u.deg,
                                               outer_radius=0.0005 * u.deg)
         # Masked subset.
-        my_reg_sky_3 = PolygonPixelRegion(vertices=PixCoord(x=[1, 1, 3, 3, 1], y=[1, 3, 3, 1, 1]))
+        my_reg_poly_1 = PolygonPixelRegion(vertices=PixCoord(x=[1, 1, 3, 3, 1], y=[1, 3, 3, 1, 1]))
+
         # Add them all.
-        bad_regions = self.subset_plugin.import_region([my_reg_sky_1, my_reg_sky_2, my_reg_sky_3],
-                                                       return_bad_regions=True,
-                                                       combination_mode='new')
+        bad_regions = self.subset_plugin.import_region(
+            [my_reg_pix_1, my_reg_sky_1, my_reg_sky_2, my_reg_poly_1, my_reg_pix_2, my_reg_pix_3],
+            return_bad_regions=True, combination_mode='new')
         assert len(bad_regions) == 0
 
-        # Mimic interactive regions (after)
-        self.imviz._apply_interactive_region('bqplot:ellipse', (-2, 0), (5, 4.5))
-        self.imviz._apply_interactive_region('bqplot:rectangle', (0, 0), (10, 10))
-
-        # Check interactive regions. We do not check if the translation is correct,
+        # Check regions. We do not check if the translation is correct,
         # that check hopefully is already done in glue-astronomy.
-        # Apparently, static region ate up one number...
-        subsets = self.imviz.plugins['Subset Tools'].get_regions()
+        # Polygon becomes MaskedSubset 1, thus cannot round-trip but made Subset 4 name unavailable.
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore',
+                                    message='Regions skipped: MaskedSubset 1')
+            subsets = self.imviz.plugins['Subset Tools'].get_regions()
         assert list(subsets.keys()) == ['Subset 1', 'Subset 2', 'Subset 3', 'Subset 5', 'Subset 6'], subsets  # noqa: E501
         assert isinstance(subsets['Subset 1'], CirclePixelRegion)
         assert isinstance(subsets['Subset 2'], CirclePixelRegion)
@@ -149,7 +165,7 @@ class TestLoadRegions(BaseImviz_WCS_NoWCS, BaseRegionHandler):
         assert isinstance(subsets['Subset 5'], EllipsePixelRegion)
         assert isinstance(subsets['Subset 6'], RectanglePixelRegion)
 
-        # Check static region
+        # Polygon becomes MaskedSubset 1.
         self.verify_region_loaded('MaskedSubset 1')
 
     def test_regions_annulus_from_load_data(self):
@@ -204,7 +220,10 @@ class TestLoadRegionsFromFile(BaseRegionHandler):
         assert len(bad_regions) == 1
 
         # Will load 8/9 and 7 of that become ROIs.
-        subsets = imviz_helper.plugins['Subset Tools'].get_regions()
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore',
+                                    message='Regions skipped: MaskedSubset 1')
+            subsets = imviz_helper.plugins['Subset Tools'].get_regions()
         assert list(subsets.keys()) == ['Subset 1', 'Subset 2', 'Subset 3',
                                         'Subset 4', 'Subset 5', 'Subset 6', 'Subset 7'], subsets
 
@@ -244,10 +263,10 @@ class TestLoadRegionsFromFile(BaseRegionHandler):
 
 class TestGetRegions(BaseImviz_WCS_NoWCS):
     def test_annulus(self):
-        # Outer circle
-        self.imviz._apply_interactive_region('bqplot:truecircle', (0, 0), (9, 9))
-        # Inner circle
-        self.imviz._apply_interactive_region('bqplot:truecircle', (2, 2), (7, 7))
+        self.imviz.plugins['Subset Tools'].import_region([
+            CirclePixelRegion(center=PixCoord(x=4.5, y=4.5), radius=4.5),  # Outer circle
+            CirclePixelRegion(center=PixCoord(x=4.5, y=4.5), radius=2.5),  # Inner circle
+        ], combination_mode="new")
 
         # At this point, there should be two normal circles.
         subsets = self.imviz.plugins['Subset Tools'].get_regions()
@@ -262,6 +281,7 @@ class TestGetRegions(BaseImviz_WCS_NoWCS):
         assert ss['Subset 2'][0]['region'] == subsets['Subset 2']
 
         # Create a third subset that is an annulus.
+        self.imviz.plugins['Subset Tools'].combination_mode = "new"
         subset_groups = self.imviz.app.data_collection.subset_groups
         new_subset = subset_groups[0].subset_state & ~subset_groups[1].subset_state
         self.viewer.apply_subset_state(new_subset)
@@ -276,5 +296,5 @@ class TestGetRegions(BaseImviz_WCS_NoWCS):
         assert subsets['Subset 3'].center == PixCoord(4.5, 4.5)
 
         # Clear the regions for next test.
-        self.imviz._delete_all_regions()
+        self.imviz.app.delete_subsets()
         assert len(self.imviz.app.data_collection.subset_groups) == 0