+++ jdaviz/core/template_mixin.py
@@ -4838,6 +4838,7 @@ class Table(PluginSubcomponent):
     # NOTE: These UI features are not covered in test coverage. Plugins making use of
     # this feature should ensure test coverage for their respective tables.
     show_rowselect = Bool(False).tag(sync=True)  # Flag to enable row selection boxes
+    multiselect = Bool(True).tag(sync=True)  # toggle single or multiselect of rows
     item_key = Unicode().tag(sync=True)  # Unique field to identify row for selection
     selected_rows = List().tag(sync=True)  # List of selected rows